+++ ChangeLog.md
@@ -13,6 +13,7 @@
 * Add the __`--file-watch`__ flag to auto-rebuild on file changes [#113](https://github.com/commercialhaskell/stack/issues/113)
 * Rename `stack docker exec` to `stack exec --plain`
 * Add the `--skip-msys` flag [#377](https://github.com/commercialhaskell/stack/issues/377)
+* `--keep-going`, turned on by default for tests and benchmarks [#478](https://github.com/commercialhaskell/stack/issues/478)
 
 ## 0.1.1.0